Move trending and solr_updater into new dir: scripts/solr_updater
