+++ docker/ol-solr-updater-start.sh
@@ -9,7 +9,7 @@ wget --output-document=$OSP_DUMP_LOCATION \
     https://archive.org/download/2023_openlibrary_osp_counts/osp_totals.db
 
 ls -la /solr-updater-data/
-python scripts/solr_updater.py $OL_CONFIG \
+python scripts/solr_updater/solr_updater.py $OL_CONFIG \
     --state-file /solr-updater-data/$STATE_FILE \
     --ol-url "$OL_URL" \
     --osp-dump "$OSP_DUMP_LOCATION" \
