+++ scripts/solr_updater/trending_updater_daily.py
