+++ scripts/solr_updater/trending_updater_hourly.py
