+++ scripts/solr_updater/trending_updater_init.py
